requirements.txt
Flask-Login==0.2.11
Flask-Principal==0.4.0
Flask-Restful==0.2.12
Flask-SQLAlchemy==1.0
Flask-Script==2.0.5
Flask-WTF==0.10.0
Flask-paginate==0.2.1
Pillow==2.5.1
Twisted==12.2.0
apns-clerk==0.1.1
chardet==2.2.1
facebook-sdk==0.4.0
flask-admin==1.0.8
flask-babelex==0.9.2
flask==0.10.1
gcm-client==0.1.4
phonenumbers==7.0.2
psycopg2==2.4
py-bcrypt==0.3
pyasn1==0.1.7
python-dateutil==1.5
pytz==2014.10
requests==2.3.0
uwsgi==2.0.6
simplejson==3.6.5
mandrill==1.0.57
bpython==0.13.1
uwsgitop==0.8
Flask-Cache==0.13.1
Flask-DebugToolbar==0.9.2
gevent==1.0.1
Flask-RQ==0.2